+++ README.md
@@ -1,11 +1,11 @@
-# Housou (放送)
+# Musen (無線)
 
-[![latest version](https://img.shields.io/github/v/release/ichiro-its/housou)](https://github.com/ichiro-its/housou/releases/)
-[![commit activity](https://img.shields.io/github/commit-activity/m/ichiro-its/housou)](https://github.com/ichiro-its/housou/commits/master)
-[![license](https://img.shields.io/github/license/ichiro-its/housou)](./LICENSE)
-[![build and test status](https://img.shields.io/github/workflow/status/ichiro-its/housou/Build%20and%20Test?label=build%20and%20test)](https://github.com/ichiro-its/housou/actions)
+[![latest version](https://img.shields.io/github/v/release/ichiro-its/musen)](https://github.com/ichiro-its/musen/releases/)
+[![commit activity](https://img.shields.io/github/commit-activity/m/ichiro-its/musen)](https://github.com/ichiro-its/musen/commits/master)
+[![license](https://img.shields.io/github/license/ichiro-its/musen)](./LICENSE)
+[![build and test status](https://img.shields.io/github/workflow/status/ichiro-its/musen/Build%20and%20Test?label=build%20and%20test)](https://github.com/ichiro-its/musen/actions)
 
-Housou (放送, broadcast) is a [ROS 2](https://docs.ros.org/en/foxy/index.html) package that provides a [UDP](https://en.wikipedia.org/wiki/User_Datagram_Protocol) broadcast communication library for a ROS 2 project.
+Musen (無線, wireless) is a [ROS 2](https://docs.ros.org/en/foxy/index.html) package that provides a [UDP](https://en.wikipedia.org/wiki/User_Datagram_Protocol) broadcast communication library for a ROS 2 project.
 This package is written in C++ and currently only works on Linux based operating system.
 In this package, the broadcast communication will be handled by a Broadcaster object that will send a message to multiple targets and a Listener object that will receive a message from multiple sources.
