+++ examples/fruits_broadcaster.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <housou/housou.hpp>
+#include <musen/musen.hpp>
 
 #include <unistd.h>
 
@@ -29,7 +29,7 @@
 
 int main()
 {
-  housou::StringBroadcaster broadcaster(8080);
+  musen::StringBroadcaster broadcaster(8080);
 
   broadcaster.enable_broadcast(false);
   broadcaster.add_target_host("localhost");
