+++ examples/position_broadcaster.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <housou/housou.hpp>
+#include <musen/musen.hpp>
 
 #include <stdlib.h>
 #include <unistd.h>
@@ -35,7 +35,7 @@ struct Position
 
 int main()
 {
-  housou::Broadcaster<Position> broadcaster(8080);
+  musen::Broadcaster<Position> broadcaster(8080);
 
   broadcaster.enable_broadcast(false);
   broadcaster.add_target_host("localhost");
