+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>housou</name>
+  <name>musen</name>
   <version>1.0.0</version>
   <description>A UDP broadcast communication library</description>
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
